Removed tokenization

+++ src/build_wikireading.py
@@ -1,40 +1,68 @@
 import logging
 import multiprocessing
+import sys
 import time
-from functools import partial
 
 from natural.date import compress
 from pymongo import MongoClient
 
 import config
-from vocabs import load_vocab
+from sling_tokenizer import SlingTokenizer
 from utils import find_full_matches, find_matches, get_chunks
 
+tokenizer = SlingTokenizer()
+
+
+def tokenize(document):
+    article_text = document['text'].strip()
+    tokens, break_levels, _ = tokenizer.tokenize(article_text)
+    document['string_sequence'] = tokens
+    document['break_levels'] = break_levels
+    document['sentence_breaks'] = [i for i, brk in enumerate(break_levels) if brk >= 3]
+    document['paragraph_breaks'] = [i for i, brk in enumerate(break_levels) if brk == 4]
+
+    tokens, _, _ = tokenizer.tokenize(document['label'])
+    document['label_sequence'] = tokens
+
+    for prop in document['properties']:
+        tokens, _, _ = tokenizer.tokenize(document['properties'][prop]['label'])
+        document['properties'][prop]['label_sequence'] = tokens
+
+    for prop in document['facts']:
+        for fact in document['facts'][prop]:
+            tokens, _, _ = tokenizer.tokenize(fact['value'])
+            if len(tokens) == 0:
+                tokens = fact['value']
+            fact['value_sequence'] = tokens
+
+    return document
 
 def build(limit):
     client = MongoClient(config.MONGO_IP, config.MONGO_PORT)
     db = client[config.DB]
     wikimerge = db[config.WIKIMERGE_COLLECTION]
+    wikireading = db['test_wikireading']
 
-    answer_vocab = load_vocab(config.ANSWER_VOCAB_PATH)
-    document_vocab = load_vocab(config.DOCUMENT_VOCAB_PATH)
-    raw_answers_vocab = load_vocab(config.RAW_ANSWER_VOCAB_PATH)
-    # type_vocab = load_vocab(config.TYPE_VOCAB_PATH)
+    i = 0
+    processed = []
+    start_time = time.time()
+    for pa in wikimerge.find({"id": {"$gte": limit[0], "$lte": limit[1]}}, {"_id": 0}):
+        text = pa['text'].strip()
+        if not text:
+            continue
 
-    for page in wikimerge.find({"id": {"$gte": limit[0], "$lte": limit[1]}}, {"_id": 0}):
-        wikireading_doc = {"key": page['id'], "break_levels": page['break_levels'],
-                           "string_sequence": page['string_sequence'], "paragraph_breaks": page['paragraph_breaks'],
-                           "sentence_breaks": page['sentence_breaks'], "text": page['text']}
-
-        # TODO Add POS Source
-        # pos_sequence = poscoll.find_one({"id": page['id']}, {"_id": 0})
-        # wikireading_doc['type_sequence'] = [type_vocab[pos] for pos in pos_sequence]
-        wikireading_doc['document_sequence'] = [document_vocab[word] for word in wikireading_doc['string_sequence']]
+        page = tokenize(pa)
+        i += 1
+        wikireading_doc = {'answer_breaks': [], 'answer_ids': ['IDs'], 'answer_location': [],
+                           'answer_sequence': ['IDs'], 'answer_string_sequence': [], 'break_levels': [],
+                           'document_sequence': ['IDs'], 'full_match_answer_location': [],
+                           'paragraph_breaks': page['paragraph_breaks'], 'question_sequence': ['IDs'],
+                           'question_string_sequence': [],
+                           'raw_answer_ids': ['IDs'], 'raw_answers': '', 'sentence_breaks': page['sentence_breaks'],
+                           'string_sequence': page['string_sequence'], 'type_sequence': ['IDs']}
 
         for prop in page['facts']:
             question = page['properties'][prop]
-            wikireading_doc['question_string_sequence'] = question['label_sequence']
-            wikireading_doc['question_sequence'] = [document_vocab[token] for token in question['label_sequence']]
             answer_string_sequence = []
             answer_breaks = []
             raw_answers = []
@@ -44,38 +72,62 @@ def build(limit):
                 if answer_string_sequence:
                     answer_breaks.append(len(answer_string_sequence))
                 raw_answers.append(fact['value'])
-                answer_sequence = fact['value_sequence']
-                answer_string_sequence += answer_sequence
-                full_match_answer_location.append(
-                    find_full_matches(wikireading_doc["string_sequence"], answer_sequence))
-                answer_location.append(find_matches(wikireading_doc["string_sequence"], answer_sequence))
+                answer = fact['value_sequence']
+                answer_string_sequence += answer
+                full_match_answer_location += find_full_matches(wikireading_doc["string_sequence"], answer)
+                answer_location += find_matches(wikireading_doc["string_sequence"], answer)
 
-            wikireading_doc['answer_sequence'] = [answer_vocab[token] for token in answer_string_sequence]
-            wikireading_doc['answer_string_sequence'] = answer_string_sequence
-            wikireading_doc['raw_answers'] = [raw_answers_vocab[token] for token in raw_answers]
-            wikireading_doc['raw_answer_ids'] = raw_answers
             wikireading_doc['answer_breaks'] = answer_breaks
+            wikireading_doc['answer_location'] = answer_location
+            wikireading_doc['answer_string_sequence'] = answer_string_sequence
             wikireading_doc['full_match_answer_location'] = full_match_answer_location
+            wikireading_doc['question_string_sequence'] = question['label_sequence']
+
+            processed.append(wikireading_doc)
+
+        if len(processed) > 500:
+            wikireading.insert_many(processed, ordered=False, bypass_document_validation=True)
+            processed = []
+
+    if processed:
+        wikireading.insert_many(processed, ordered=False, bypass_document_validation=True)
+
+    elapsed = int(time.time() - start_time)
+    return i, elapsed
 
 
-def make_wikireading(configs):
+def wikireading():
     client = MongoClient(config.MONGO_IP, config.MONGO_PORT)
     db = client[config.DB]
-    wikimerge = db[config.WIKIMERGE_COLLECTION]
-    documents_id = list(wikimerge.find({}, {"id": 1, "_id": 0}).sort("id"))
-    client.close()
+    wikipedia = db[config.WIKIPEDIA_COLLECTION]
+    wikidocs = list(wikipedia.find({}, {'wikidata_id': 1, '_id': 0}).sort('wikidata_id'))
+    chunks = get_chunks(wikidocs, config.CHUNK_SIZE, 'wikidata_id')
+    del wikidocs
     start_time = time.time()
     total = 0
-    pool = multiprocessing.Pool(config.NUM_WORKERS)
-    chunks = get_chunks(documents_id, config.CHUNK_SIZE, 'wikidata_id')
-    for n, elapsed in pool.map(partial(build, configs=configs), chunks):
-        total += n
-        part = int(time.time() - start_time)
-        logging.info("Processed {} ({} in total) documents in {} (running time {})".format(n, total,
-                                                                                           compress(elapsed),
-                                                                                           compress(part)))
 
-    pool.terminate()
+    if config.NUM_WORKERS == 1:
+        for chunk in chunks:
+            build(chunk)
+    else:
+        chunks = [(x.decode(), y.decode()) for x, y in chunks]
+        pool = multiprocessing.Pool(config.NUM_WORKERS)
+        for n, elapsed in pool.imap(build, chunks):
+            total += n
+            part = int(time.time() - start_time)
+            logging.info(
+                "Processed {} ({} in total) documents in {} (running time {})".format(n, total, compress(elapsed),
+                                                                                      compress(part)))
+
+        pool.terminate()
     elapsed = int(time.time() - start_time)
     logging.info("Processed {} documents in {}".format(total, compress(elapsed)))
     return
+
+
+if __name__ == '__main__':
+    logging.basicConfig(format='%(asctime)s - %(module)s - %(levelname)s - %(message)s', level=logging.INFO)
+
+    logging.info("Running %s", " ".join(sys.argv))
+    wikireading()
+    logging.info("Completed %s", " ".join(sys.argv))
